scripts: Add script to rename (add label to) release assets
The only way I've found to add a label to existing release assets on
GitHub is via the GitHub API. Add a script to "rename" (add a label) to
existing assets.

In particular, we rename the archive containing the sources for snoop
and its submodules.

After running the script, instead of being listed as
"retsnoop-all-sources-<tag>.tar.gz", the asset appears on GitHub's
interface as "Full source code with submodules (tar.gz)". More assets
could easily be renamed as well by extending the array in the script.

The script is not run from CI, because it requires the release to be
published (we only automated the creation of the draft release).

The script requires the GitHub command line ("gh") to be set up. If the
asset is found but the update fails with a 404 error, this is usually
synonymous with insufficient permissions for the GitHub token in use.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>